Add proper copyright attributions as suggested by FSF.

+++ gnubg/board3d/Occluders.c
@@ -1,3 +1,21 @@
+/*
+ * Copyright (C) 2019 Jon Kinsey <jonkinsey@gmail.com>
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ *
+ * $Id$
+ */
 
 #include "config.h"
 #include "legacyGLinc.h"
@@ -131,7 +149,7 @@ updateMovingPieceOccPos(const BoardData* bd, BoardData3d* bd3d)
 			float id[4][4];
 
 			bd3d->Occluders[LAST_PIECE].rotator = 1;
-			bd3d->Occluders[LAST_PIECE].rot[1] = -90 * bd3d->rotateMovingPiece * bd->turn;
+			bd3d->Occluders[LAST_PIECE].rot[1] = -90 * bd3d->rotateMovingPiece * (float) bd->turn;
 			makeInverseTransposeMatrix(id, bd3d->Occluders[LAST_PIECE].trans);
 			makeInverseRotateMatrixX(bd3d->Occluders[LAST_PIECE].invMat, bd3d->Occluders[LAST_PIECE].rot[1]);
 			matrixmult(bd3d->Occluders[LAST_PIECE].invMat, (ConstMatrix)id);
